+++ lib/list.h
@@ -23,7 +23,7 @@
 #ifndef _LIST_H
 #define _LIST_H
 
-#include <stddef.h>
+#include <sys/types.h>
 
 /* list definition */
 typedef struct _element *element;
@@ -57,11 +57,11 @@ extern list alloc_list(void (*free_func) (void *), void (*dump_func) (void *));
 extern void free_list(list *);
 extern void free_list_elements(list l);
 extern void free_list_element(list l, element e);
-extern void *list_element(list l, unsigned num);
+extern void *list_element(list l, size_t num);
 extern void dump_list(list l);
 extern void list_add(list l, void *data);
 extern void list_del(list l, void *data);
 extern list alloc_mlist(void (*free_func) (void *), void (*dump_func) (void *), size_t size);
-extern void free_mlist(list l, int size);
+extern void free_mlist(list l, size_t size);
 
 #endif
